skriva om koden med en useState
