+++ src/types/types.ts
@@ -1,5 +1,3 @@
-
-
 export enum BoardProps {
   Rows = 6,
   Cols = 7,
@@ -7,17 +5,10 @@ export enum BoardProps {
 
 export type Matrix = (string | null)[][];
 
-
 export type Move = {
   row: number;
   col: number;
 };
-
-export enum PlayerType {
-  Player = 1,
-  Easy,
-  Hard,
-}
 
 export enum Color {
   Red = 1,
@@ -29,23 +20,35 @@ export type ColumnStatus = (0 | 1 | 2 | 3 | 4 | 5 | 6)[];
 export interface BoardState {
   matrix: (string | null)[][];
   currentPlayerColor: string;
-  
 }
 
-export enum RegExes {
-  Column = 0,
-  GameType = 1,
-  Level = 2,
-  PlayerName = 3,
+// export enum RegExes {
+//   Column = 0,
+//   GameType = 1,
+//   Level = 2,
+//   PlayerName = 3,
+// }
+// export type GamePlayer = {
+//   name: string;
+//   playerType: PlayerType;
+// };
+
+export enum PlayerType {
+  Player = 1,
+  Easy,
+  Hard,
 }
+
 export enum GameType {
   Dual = 1,
   Single,
   AI,
 }
-export type GamePlayer = {
-  name: string;
-  playerType: PlayerType;
+
+export type Setup = {
+  playerOneName: string;
+  playerTwoName: string;
+  playerOneType: PlayerType;
+  playerTwoType: PlayerType;
+  gameType: GameType;
 };
-
-
